# config/swagger.yaml
openapi: 3.0.1
info:
  title: npsstore API
  description: This is a Backend API of NBSStore
  version: 1.0.0
  termsOfService: http://quocbaoit.com/terms/
  contact:
    email: NBSstoreProject@quocbaoit.com
...
paths:
...
  /api/v1/order:
    post:
      summary: Create new order
      description: Create new order
      tags:
        - Order
      security:
        - bearerAuth: []
      requestBody:
        description: Order form
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderForm'
      responses:
        '400':
          description: No order items/One or more product order quantity exceed available quantity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: Users get their orders
      description: Users get their orders
      tags:
        - Order
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/order/all:
    get:
      summary: Get all orders
      description: Get all orders
      tags:
        - Order
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - type: object
                      properties:
                        shippingAddress:
                          $ref: '#/components/schemas/ShippingAddress'
                    - type: object
                      properties:
                        user:
                          $ref: '#/components/schemas/OrderedUser'
                    - type: object
                      properties:
                        orderItems:
                          type: array
                          items:
                            $ref: '#/components/schemas/OrderItemWithID'
                    - type: object
                      properties:
                        _id: 
                          type: string
                          description: Order ID
                        patmentMethod:
                          type: string
                          description: Payment method
                          default: "Paypal"
                        taxPrice:
                          type: number
                          description: Tax price
                          default: 0.0
                        shippingPrice:
                          type: number
                          description: Shipping price
                          default: 0.0         
                        totalPrice:
                          type: number
                          description: Total price
                          default: 0.0 
                        isPaid:
                          type: boolean
                          description: Is paid yet ?
                          default: 0.0 
                    
        '400':
          description: No order items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/order/{id}:
    get:
      summary: Get order by id
      description: Get order by id
      tags:
        - Order
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Order ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete an order
      description: Delete an order
      security:
        - bearerAuth: []
      tags:
        - Order
      parameters:
        - name: id
          in: path
          description: Order ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: 
                    type: string
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/order/{id}/pay:
    patch:
      summary: Confirm order is paid
      description: Confirm order is pad
      tags:
        - Order
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Order ID
          required: true
          schema:
            type: string
      requestBody:
        description: Payload to confirm an order is paid
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmPaidForm'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error' 
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/v1/order/{id}/disable:
    patch:
      summary: Disable an order
      description: Disable an order
      security:
        - bearerAuth: []
      tags:
        - Order
      parameters:
        - name: id
          in: path
          description: Order ID
          required: true
          schema:
            type: string
      requestBody:
        description: isDisabled value
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                isDisabled:
                  type: boolean
                  description: Is disabled or not
                  nullable: false
                  default: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/order/{id}/restore:
    patch:
      summary: Restore disabled order
      description: Restore disabled order
      security:
        - bearerAuth: []
      tags:
        - Order
      parameters:
        - name: id
          in: path
          description: Order ID
          required: true
          schema:
            type: string
      requestBody:
        description: isDisabled value
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                isDisabled:
                  type: boolean
                  description: Is disabled or not
                  nullable: false
                  default: false
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error' 
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        message:
          type: string
          description: error message
        stack:
          type: string
          description: error message description
...
    Timestamp:
      type: object
      properties:
        createAt:
          type: string
          description: The date and time when this record is created
          format: date-time
        updatedAt:
          type: string
          description: The date and time when this record is updated
          format: date-time
...
    OrderItem:
      type: object
      properties:
        name:
          type: string
          description: Order item name
        qty:
          type: number
          description: Order item quantity
        image:
          type: string
          description: Order item image's url
        price:
          type: number
          description: Order item price
        product:
          type: string
          description: Order item ID
    ShippingAddress:
      type: object
      properties:
        address:
          type: string
          description: Shipping address
        city:
          type: string
          description: Shipping city
        postalCode:
          type: string
          description: Shipping postal code
        country:
          type: string
          description: Shipping country
...
    OrderForm:
      type: object
      properties:
        orderItems:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        shippingAddress:
          $ref: '#/components/schemas/ShippingAddress'
        paymentMethod:
          type: string
          description: Payment method
          default: Paypal
        itemsPrice:
          type: number
          description: Items price
        taxPrice:
          type: number
          description: Tax price
          default: 0
        shippingPrice:
          type: number
          description: Shipping price
          default: 0
        totalPrice:
          type: number
          description: Total price
          default: 0
    Order:
      type: object
      properties:
        user:
          type: string
          description: Ordered user ID
        orderItems:
          type: array
          items:
            allOf:
              - $ref: '#/components/schemas/OrderItem'
              - type: object
                properties:
                  _id:
                    type: string
                    description: Order item ID
        shippingAddress:
          $ref: '#/components/schemas/ShippingAddress'
        paymentMethod:
          type: string
          description: Payment method
          default: Paypal
        taxPrice:
          type: number
          description: Tax price
          default: 0
        shippingPrice:
          type: number
          description: Shipping price
          default: 0
        totalPrice:
          type: number
          description: Total price
          default: 0     
        isPaid:
          type: boolean
          description: Is paid yet ?
        isDelivered:
          type: boolean
          description: Is deliverd yet ?
        _id: 
          type: string
          description: Order ID
        timestamp: 
          $ref: '#/components/schemas/Timestamp'
        __v: 
          type: number
          description: Document version key
        isDisabled:
          type: boolean
          description: is Disabled or not
    OrderedUser:
      type: object
      properties:
        _id:
          type: string
          description: User ID
        name:
          type: string
          description: User name
    OrderItemWithID:
      allOf:
        - $ref: '#/components/schemas/OrderItem'
        - type: object
          properties:
            _id:
              type: string
              description: Order item ID
    ConfirmPaidForm:
      type: object
      properties:
        id:
          type: string
          description: Ordered user ID
        status:
          type: string
          description: Paid status
        update_time: 
          type: string
          description: The date and time when this record is updated
          format: date-time
        email_address:
          type: string
          description: Ordered user address
...
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: apiKey
      description: "<i>JWT Authorization header using the Bearer scheme.</i> <b>Example</b>: <i><u>Bearer 12345abcdef</u></i>"
      in: header          # can be "header", "query" or "cookie"
      name: Authorization # name of the header, query parameter or cookie
# 2) Apply the security globally to all operations
security:
  - bearerAuth: [] # use the same name as above